Add sitegeist
sitegeist is a German web agency based in Hamburg. Besides allowing and encouraging their employees to contribute to and/or publish their own open source projects, the company also supports the communities and development teams behind those open source products they are using on a regular basis. For the larger part these are:

* TYPO3 (https://typo3.org/)
* Neos CMS (https://www.neos.io/)

Their support for these projects includes:

* Financial support
* Organization and sponsoring of conferences, meetups and other event formats
* Contribution in community building and software development